GEODE-6176: Make FunctionService testable with internal delegates
Introduce new internal FunctionExecutionService API interfaces:
* FunctionExecutionService
* InternalFunctionExecutionService

Implement interfaces with:
* InternalFunctionExecutionServiceImpl

Collapse FunctionServiceManager into InternalFunctionExecutionServiceImpl

Change the two static API classes to delegate to an instance of
InternalFunctionExecutionServiceImpl:
* FunctionService (User API) -- only has private internal changes
* InternalFunctionService -- now extends FunctionService

Geode classes that use the static API classes directly can now be
changed to depend on one of the two interfaces using constructor
injection to facilitate unit testing without PowerMock.
