+++ geode-core/src/main/java/org/apache/geode/internal/cache/properties.html
@@ -1691,7 +1691,7 @@ Allow query on region with heterogeneous objects
 <p>
 <em>Boolean</em> (defaults to false)
 <p>
-See <code>org.apache.geode.cache.execute.internal.FunctionServiceManager.#RANDOM_onMember</code>
+See <code>org.apache.geode.internal.cache.execute.InternalFunctionExecutionServiceImpl.#RANDOM_onMember</code>
 <p>
 When set, onMember execution will be executed on a random member.
 </dd>
